refactor(server): move notification constants to models
