+++ server/app/src/main/kotlin/pos/ambrosia/models/WebPushConstants.kt
@@ -1,0 +1,10 @@
+package pos.ambrosia.models
+
+object WebPushPayloadTypes {
+    const val ADMIN_ACTIVITY = "admin_activity"
+}
+
+object WebPushStatusCodes {
+    const val NOT_FOUND = 404
+    const val GONE = 410
+}
